// amplify/functions/commons/types/AuctionListingFileTypes.ts
export interface AuctionListingData {
  IMAGE1: string;
  IMAGE2?: string;
  IMAGE3?: string;
  IMAGE4?: string;
  IMAGE5?: string;
  IMAGE6?: string;
  TITLE: string;
  SHORT_TITLE?: string;
  CATEGORY1: string;
  CATEGORY2?: string;
  CATEGORY3?: string;
  SUBCATEGORY1: string;
  SUBCATEGORY2?: string;
  SUBCATEGORY3?: string;
  SUBCATEGORY4?: string;
  SUBCATEGORY5?: string;
  DESCRIPTION?: string;
  LOT_TYPE: string;
  COSMETIC_CONDITION?: string;
  RESALE_REQUIREMENTS?: string;
  ACCESSORIES?: string;
  INSPECTION_STATUS?: string;
  TOTAL_EX_RETAIL_PRICE: string;
  TOTAL_UNITS: string;
  AVG_COST_PER_UNIT?: string;
  SELLER_NOTES?: string;
  SHIPPING_TYPE: string;
  SHIPPING_COST?: string;
  WAREHOUSE_LOCATION_ADDRESS1: string;
  WAREHOUSE_LOCATION_ADDRESS2?: string;
  WAREHOUSE_LOCATION_ADDRESS3?: string;
  WAREHOUSE_LOCATION_CITY: string;
  WAREHOUSE_LOCATION_STATE: string;
  WAREHOUSE_LOCATION_STATE_CODE: string;
  WAREHOUSE_LOCATION_COUNTRY: string;
  WAREHOUSE_LOCATION_COUNTRY_CODE: string;
  WAREHOUSE_LOCATION_ZIPCODE: string;
  SHIP_FROM_LOCATION?: string;
  FREIGHT_TYPE: string;
  PIECE_COUNT: string;
  WEIGHT_TYPE: string;
  ESTIMATED_WEIGHT: string;
  REFRIGERATED?: string;
  FDA_REGISTERED: string;
  HAZMAT: string;
  LOT_PACKAGING: string;
  NUMBER_OF_PALLETS: string;
  PALLET_SPACES: string;
  NUMBER_OF_TRUCKLOADS: string;
  NUMBER_OF_SHIPMENTS: string;
  SHIPPING_NOTES?: string;
  ADDITIONAL_INFORMATION?: string;
  BIDDING_REQUIREMENTS?: string;
}

export interface ManifestData {
  LOT_ID?: string;
  PALLET_ID?: string;
  BRAND: string;
  DEPARTMENT?: string;
  CATEGORY: string;
  SUBCATEGORY: string;
  TITLE: string;
  UPC: string;
  ASIN?: string;
  MODEL?: string;
  SKU: string;
  ITEM_DESCRIPTION: string;
  MPN?: string;
  UNIT_QTY: string;
  UNIT_RETAIL: string;
  UNIT_WEIGHT_TYPE?: string;
  UNIT_WEIGHT?: string;
  UNIT_DIMENSION_TYPE?: string;
  UNIT_LENGTH?: string;
  UNIT_WIDTH?: string;
  UNIT_HEIGHT?: string;
  CASE_PACK?: string;
  CASE_NET_WEIGHT_TYPE?: string;
  CASE_NET_WEIGHT?: string;
  CASE_DIMENSION_TYPE?: string;
  CASE_LENGTH?: string;
  CASE_WIDTH?: string;
  CASE_HEIGHT?: string;
  CONDITION: string;
  COSMETIC_CONDITION?: string;
  PACKAGING?: string;
  ACCESSORIES?: string;
  HAZMAT: string;
}